backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
flower==2.0.1

openai>=1.0.0
python-telegram-bot==20.8
httpx>=0.25.2
python-dotenv==1.0.1
boto3==1.34.23
requests==2.31.0
beautifulsoup4==4.12.3

aiosqlite
google-generativeai
